Process rows for transform DB operation
This patch adds a framwwork for getting datatypes from a single
row of flatbuffer. Along with that this patch adds to the basic layout
needed for transform DB opertion

a. Adds the option of --start-obj
b. Adds the Apache Arrow dependency in the install_deps.sh
c. Links the cls_tabular with arrow libraries