Refactor the CI pipeline
Move the pipeline out from `ci/wf.yml` to the default popper workflow
file (`.popper.yml`) on the root of the project.

The ci/ folder is now named docker/ as it only contains Dockerfiles now.
The builder image definition has been updated so it doesn't use the
`builder_entrypoint.sh` script and instead use the parent's entrypoint
(bash). This script is now in `scripts/build.sh`.

In this new `scripts/` folder there's a `run_tests.sh and `micro-osd.sh`
files that are used by test step of the pipeline.

The CMakeLists.txt file was moved to src/

The README and travis config file have been updated accordingly.