Adds cmake config for sdk cls development
Adds Dockerfile with dependencies for building cls classes. The base
image is the upstream ceph image. With gcc, make, cmake, etc. on top, as
well as the RADOS SDK.

This also includes the SDK example from upstream Ceph to show how it
works.

A minimal popper pipeline is also included to automate the build step.